LPS-164308 Extract Conditions from ObjectValidationTabs to its own component
