+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/ObjectValidation/useObjectValidationForm.d.ts
@@ -12,8 +12,8 @@
  * details.
  */
 
-import {ChangeEventHandler} from 'react';
 import {FormError} from '@liferay/object-js-components-web';
+import {ChangeEventHandler} from 'react';
 interface IUseObjectValidationForm {
 	initialValues: Partial<ObjectValidation>;
 	onSubmit: (validation: ObjectValidation) => void;
